[Windows GUI] Sync with MAMEUI
git-svn-id: svn://messdev.no-ip.org/mess@1679 9849079f-6229-0410-a07f-84406ce305ba

+++ src/osd/winui/datamap.c
@@ -719,7 +719,7 @@ static void populate_control(datamap *map, HWND control, core_options *opts, dat
 						entry->use_trackbar_options = TRUE;
 						entry->trackbar_min = minval_float;
 						entry->trackbar_max = maxval_float;
-						entry->trackbar_increments = 0.05f;
+						entry->trackbar_increments = (float)0.05;
 						break;
 				}
 			}
